feat(Titlebar): Component renamed
